pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lelouch"
version = "1.0.0"
description = "Agentic web video scraper - describe footage in natural language, get clips automatically"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "litellm>=1.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "ddgs>=7.0.0",
    "opencv-python>=4.8.0",
    "yt-dlp>=2024.0.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]

[project.scripts]
lelouch = "lelouch.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
